requirement.txt
Flask
mysql-connector-python 
werkzeug 
pandas
Nginx
Gunicorn